refactor, review
