+++ projects/batfish/src/test/java/org/batfish/job/ConvertConfigurationJobTest.java
@@ -1,41 +1,30 @@
 package org.batfish.job;
 
-import static org.batfish.datamodel.Configuration.DEFAULT_VRF_NAME;
-import static org.batfish.datamodel.Names.generatedTenantVniInterfaceName;
-import static org.batfish.job.ConvertConfigurationJob.addTenantVniInterfaces;
 import static org.batfish.job.ConvertConfigurationJob.finalizeConfiguration;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasKey;
 import static org.junit.Assert.assertThat;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
 import org.batfish.common.VendorConversionException;
 import org.batfish.common.Warnings;
 import org.batfish.datamodel.AclAclLine;
-import org.batfish.datamodel.BumTransportMethod;
 import org.batfish.datamodel.Configuration;
 import org.batfish.datamodel.ConfigurationFormat;
-import org.batfish.datamodel.EmptyIpSpace;
 import org.batfish.datamodel.ExprAclLine;
 import org.batfish.datamodel.HeaderSpace;
 import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.InterfaceType;
-import org.batfish.datamodel.Ip;
 import org.batfish.datamodel.IpAccessList;
 import org.batfish.datamodel.IpSpaceReference;
 import org.batfish.datamodel.LineAction;
-import org.batfish.datamodel.Vrf;
 import org.batfish.datamodel.acl.AndMatchExpr;
 import org.batfish.datamodel.acl.MatchHeaderSpace;
 import org.batfish.datamodel.acl.NotMatchExpr;
 import org.batfish.datamodel.acl.OrMatchExpr;
 import org.batfish.datamodel.routing_policy.communities.CommunityMatchExprReference;
 import org.batfish.datamodel.transformation.Transformation;
-import org.batfish.datamodel.vxlan.Layer3Vni;
 import org.batfish.job.ConvertConfigurationJob.CollectIpSpaceReferences;
 import org.junit.Rule;
 import org.junit.Test;
@@ -142,42 +131,5 @@ public final class ConvertConfigurationJobTest {
     _thrown.expect(VendorConversionException.class);
     _thrown.expectMessage(containsString("Undefined reference"));
     finalizeConfiguration(c, new Warnings());
-  }
-
-  @Test
-  public void testAddTenantVniInterfaces() {
-    Configuration c = new Configuration("c", ConfigurationFormat.CISCO_NX);
-    c.setDefaultCrossZoneAction(LineAction.PERMIT);
-    c.setDefaultInboundAction(LineAction.PERMIT);
-    Vrf.Builder vb = Vrf.builder().setOwner(c);
-    Vrf v1 = vb.setName("v1").build();
-    Vrf v2 = vb.setName("v2").build();
-    Ip sourceAddress = Ip.parse("10.0.0.1");
-    Layer3Vni.Builder l3b =
-        Layer3Vni.builder()
-            .setBumTransportIps(ImmutableSet.of())
-            .setBumTransportMethod(BumTransportMethod.UNICAST_FLOOD_GROUP)
-            .setSrcVrf(DEFAULT_VRF_NAME)
-            .setUdpPort(5);
-    v1.addLayer3Vni(l3b.setVni(1).setSourceAddress(sourceAddress).build());
-    v2.addLayer3Vni(l3b.setVni(2).setSourceAddress(null).build());
-    String vni1IfaceName = generatedTenantVniInterfaceName(1);
-    String vni2IfaceName = generatedTenantVniInterfaceName(2);
-    addTenantVniInterfaces(c);
-
-    // v1 vni 1
-    assertThat(c.getAllInterfaces(v1.getName()), hasKey(vni1IfaceName));
-
-    Interface vni1Iface = c.getAllInterfaces().get(vni1IfaceName);
-
-    assertThat(vni1Iface.getAdditionalArpIps(), equalTo(sourceAddress.toIpSpace()));
-
-    // v2 vni 2
-
-    assertThat(c.getAllInterfaces(v2.getName()), hasKey(vni2IfaceName));
-
-    Interface vni2Iface = c.getAllInterfaces().get(vni2IfaceName);
-
-    assertThat(vni2Iface.getAdditionalArpIps(), equalTo(EmptyIpSpace.INSTANCE));
   }
 }
